[AI4DSOC] Remove Assistant icon from row action in alert summary table (#219141)
## Summary

During testing we discovered an issue related to the Assistant icon
displayed in the AI4DSOC alert summary page table.

The issue is related to the fact that the alert registers its context
with the assistant. When the assistant flyout is opened from the `Ask AI
assistant` button in the alert flyout, we have one flyout which means
one alert in the assistant. But with the Assistant icon shown on each
row of the table, we need to register context for each alert. That means
that x alert buttons equals to x alerts in the assistant.
This is how it translates in the UI:

![Screenshot 2025-04-24 at 12 54
50 PM](https://github.com/user-attachments/assets/d59b1cdd-ceef-499c-ba8c-55af38ea6bac)

Redesigning how the assistant context works is very involved and too
risky to do at such a late time.

This PR removes the Assistant button from the row actions, which solves
the issue:

https://github.com/user-attachments/assets/0fd94cc1-4fbd-4e70-9790-22e4913477ff

Also, @stephmilovic noticed that when the assistant is opened from a
different launch point than the suggested prompt, and the suggested
prompt is still in the DOM, its prompt context shows up empty:
<img width="600" alt="Screenshot 2025-04-24 at 11 41 23 AM"
src="https://github.com/user-attachments/assets/0ceb3ffc-72e5-425c-b550-9b8d5896f359"
/>
She resolved this by adding a check in the ContextPills component for
`description.length > 0`. This would be resolved by the on click issue,
so might be worth coming back to.
This change is visible in the second commit and stolen from [this
PR](https://github.com/elastic/kibana/pull/219144).

### Checklist

- [x] [Unit or functional
tests](https://www.elastic.co/guide/en/kibana/master/development-tests.html)
were updated or added to match the most common scenarios

relates to https://github.com/elastic/security-team/issues/11973

(cherry picked from commit cc98975ced24d76773459c488953748313a84dc7)