[Security Solution][AI4DSOC] Fix table not applying alert tags for Attack discovery and Cases pages in AI4DSOC (#219410)
## Summary

This PR fixes an issue with the `apply alert tags` functionality. This
[setAlertTags](https://github.com/elastic/kibana/blob/main/x-pack/solutions/security/plugins/security_solution/public/common/components/toolbar/bulk_actions/use_set_alert_tags.tsx)
hook introduced in [this
PR](https://github.com/elastic/kibana/pull/157786) was making use of a
`useRef` and a `useCallback` which lead to the table having a `null`
reference when it was expecting a `function`.

This meant that for parent table components that were written
efficiently (meaning that do not render too many times), the set alert
tags bulk actions was just plain not working. As can be seen in the
following video:

https://github.com/user-attachments/assets/50545677-d7a7-4d45-828e-89bdd7b2de34

And this is with the fix in this PR

https://github.com/user-attachments/assets/b1f07e55-43bb-4d01-bfd9-3b0dd23e30cc

The PR also makes 2 very small adjustments to the AI4DSOC alerts table
in the Cases and Attack discovery pages, to make sure that the table
refreshes after applying alert tags.

### Checklist

- [x] [Unit or functional
tests](https://www.elastic.co/guide/en/kibana/master/development-tests.html)
were updated or added to match the most common scenarios

(cherry picked from commit 6f4d780a4e1ff1459d470b74ec243156b4b0d1e7)